fixed arraylist & performance improvement

+++ src/task/jsonimportertask.cpp
@@ -484,7 +484,7 @@ void JSONImporterTask::updateMsgBox ()
             +"m "+std::to_string(diff.seconds())+"s";
 
     msg += "Elapsed Time: "+time_str+"\n";
-    msg += "Bytes read: "+String::doubleToStringPrecision(static_cast<float>(bytes_read_)*1e-6,2)+" MB\n";
+    msg += "Bytes read: "+String::doubleToStringPrecision(static_cast<double>(bytes_read_)*1e-6,2)+" MB\n";
     msg += "Objects read: "+std::to_string(objects_read_)+"\n";
     msg += "Objects parsed: "+std::to_string(objects_parsed_)+"\n";
     msg += "Objects mapped: "+std::to_string(objects_mapped_)+"\n";
